*crc:crc16 - небольшая оптимизация кода

+++ sources/crc/crc16.h
@@ -14,7 +14,7 @@ extern "C" {
 #define CRC16_GOOD	0x0000	/* Good final checksum value */
 
 unsigned short crc16 (unsigned short sum, unsigned const char *buf,
-	unsigned short len);
+	unsigned len);
 
 unsigned short crc16_byte (unsigned short sum, unsigned char byte);
